+++ src/i_video.h
@@ -95,7 +95,7 @@ void GetScreenResolution(void);
 // Takes full 8 bit values.
 void I_SetPalette(byte *playpal);
 void I_SetExternalAutomapPalette(void);
-void I_SetPaletteWithBrightness(byte *playpal, double brightness);
+void I_SetPaletteWithBrightness(byte *playpal, float brightness);
 
 void I_UpdateBlitFunc(bool shaking);
 bool I_CreateExternalAutomap(void);
